Move MergeTree related UT to mergetree module
