+++ backends-clickhouse/src/test/scala/org/apache/gluten/execution/mergetree/GlutenClickHouseMergeTreeWriteOnS3Suite.scala
@@ -14,7 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.gluten.execution
+package org.apache.gluten.execution.mergetree
+
+import org.apache.gluten.execution.{BasicScanExecTransformer, FileSourceScanExecTransformer, GlutenClickHouseTPCHAbstractSuite}
 
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.SaveMode
@@ -23,18 +25,16 @@ import org.apache.spark.sql.delta.files.TahoeFileIndex
 import org.apache.spark.sql.execution.adaptive.AdaptiveSparkPlanHelper
 import org.apache.spark.sql.execution.datasources.v2.clickhouse.metadata.AddMergeTreeParts
 
-import org.apache.commons.io.FileUtils
-import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.{FileSystem, Path}
+import _root_.org.apache.commons.io.FileUtils
+import io.minio._
+import io.minio.messages.DeleteObject
 
 import java.io.File
+import java.util
 
 import scala.concurrent.duration.DurationInt
 
-// Some sqls' line length exceeds 100
-// scalastyle:off line.size.limit
-
-class GlutenClickHouseMergeTreeWriteOnHDFSSuite
+class GlutenClickHouseMergeTreeWriteOnS3Suite
   extends GlutenClickHouseTPCHAbstractSuite
   with AdaptiveSparkPlanHelper {
 
@@ -43,6 +43,12 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
   override protected val tablesPath: String = basePath + "/tpch-data"
   override protected val tpchQueries: String = rootPath + "queries/tpch-queries-ch"
   override protected val queriesResults: String = rootPath + "mergetree-queries-output"
+
+  private val client = MinioClient
+    .builder()
+    .endpoint(MINIO_ENDPOINT)
+    .credentials(S3_ACCESS_KEY, S3_SECRET_KEY)
+    .build()
 
   override protected def createTPCHNotNullTables(): Unit = {
     createNotNullTPCHTablesInParquet(tablesPath)
@@ -56,34 +62,41 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
       .set("spark.sql.autoBroadcastJoinThreshold", "10MB")
       .set("spark.sql.adaptive.enabled", "true")
       .set("spark.gluten.sql.columnar.backend.ch.runtime_config.logger.level", "error")
-      .set(
-        "spark.gluten.sql.columnar.backend.ch.runtime_settings.mergetree.merge_after_insert",
-        "false")
-    // .set("spark.gluten.sql.columnar.backend.ch.runtime_config.path", "/data") // for local test
+      .set("spark.gluten.sql.columnar.backend.ch.runtime_config.path", "/data")
   }
 
   override protected def beforeEach(): Unit = {
     super.beforeEach()
-    val conf = new Configuration
-    conf.set("fs.defaultFS", HDFS_URL)
-    val fs = FileSystem.get(conf)
-    fs.delete(new org.apache.hadoop.fs.Path(HDFS_URL), true)
-    FileUtils.deleteDirectory(new File(HDFS_METADATA_PATH))
-    FileUtils.forceMkdir(new File(HDFS_METADATA_PATH))
+    if (client.bucketExists(BucketExistsArgs.builder().bucket(BUCKET_NAME).build())) {
+      val results =
+        client.listObjects(ListObjectsArgs.builder().bucket(BUCKET_NAME).recursive(true).build())
+      val objects = new util.LinkedList[DeleteObject]()
+      results.forEach(
+        obj => {
+          objects.add(new DeleteObject(obj.get().objectName()))
+        })
+      val removeResults = client.removeObjects(
+        RemoveObjectsArgs.builder().bucket(BUCKET_NAME).objects(objects).build())
+      removeResults.forEach(result => result.get().message())
+      client.removeBucket(RemoveBucketArgs.builder().bucket(BUCKET_NAME).build())
+    }
+    client.makeBucket(MakeBucketArgs.builder().bucket(BUCKET_NAME).build())
+    FileUtils.deleteDirectory(new File(S3_METADATA_PATH))
+    FileUtils.forceMkdir(new File(S3_METADATA_PATH))
   }
 
   override protected def afterEach(): Unit = {
     super.afterEach()
-    FileUtils.deleteDirectory(new File(HDFS_METADATA_PATH))
+    FileUtils.deleteDirectory(new File(S3_METADATA_PATH))
   }
 
   test("test mergetree table write") {
     spark.sql(s"""
-                 |DROP TABLE IF EXISTS lineitem_mergetree_hdfs;
+                 |DROP TABLE IF EXISTS lineitem_mergetree_s3;
                  |""".stripMargin)
 
     spark.sql(s"""
-                 |CREATE TABLE IF NOT EXISTS lineitem_mergetree_hdfs
+                 |CREATE TABLE IF NOT EXISTS lineitem_mergetree_s3
                  |(
                  | l_orderkey      bigint,
                  | l_partkey       bigint,
@@ -103,15 +116,15 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
                  | l_comment       string
                  |)
                  |USING clickhouse
-                 |LOCATION '$HDFS_URL/test/lineitem_mergetree_hdfs'
-                 |TBLPROPERTIES (storage_policy='__hdfs_main')
+                 |LOCATION 's3a://$BUCKET_NAME/lineitem_mergetree_s3'
+                 |TBLPROPERTIES (storage_policy='__s3_main')
                  |""".stripMargin)
 
     spark.sql(s"""
-                 | insert into table lineitem_mergetree_hdfs
+                 | insert into table lineitem_mergetree_s3
                  | select * from lineitem
                  |""".stripMargin)
-    FileUtils.deleteDirectory(new File(HDFS_METADATA_PATH))
+    FileUtils.deleteDirectory(new File(S3_METADATA_PATH))
     val sqlStr =
       s"""
          |SELECT
@@ -126,7 +139,7 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
          |    avg(l_discount) AS avg_disc,
          |    count(*) AS count_order
          |FROM
-         |    lineitem_mergetree_hdfs
+         |    lineitem_mergetree_s3
          |WHERE
          |    l_shipdate <= date'1998-09-02' - interval 1 day
          |GROUP BY
@@ -157,16 +170,62 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
         assertResult(1)(addFiles.size)
         assertResult(600572)(addFiles.head.rows)
     }
-    spark.sql("drop table lineitem_mergetree_hdfs")
+    eventually(timeout(10.seconds), interval(2.seconds)) {
+      verifyS3CompactFileExist("lineitem_mergetree_s3")
+    }
+    spark.sql("drop table lineitem_mergetree_s3") // clean up
+  }
+
+  private def verifyS3CompactFileExist(table: String): Unit = {
+    val args = ListObjectsArgs
+      .builder()
+      .bucket(BUCKET_NAME)
+      .recursive(true)
+      .prefix(table)
+      .build()
+    var objectCount: Int = 0
+    var metadataGlutenExist: Boolean = false
+    var metadataBinExist: Boolean = false
+    var dataBinExist: Boolean = false
+    var hasCommits = false
+    client
+      .listObjects(args)
+      .forEach(
+        obj => {
+          objectCount += 1
+          val objectName = obj.get().objectName()
+          if (objectName.contains("metadata.gluten")) {
+            metadataGlutenExist = true
+          } else if (objectName.contains("meta.bin")) {
+            metadataBinExist = true
+          } else if (objectName.contains("data.bin")) {
+            dataBinExist = true
+          } else if (objectName.contains("_commits")) {
+            // Spark 35 has _commits directory
+            // table/_delta_log/_commits/
+            hasCommits = true
+          }
+        })
+
+    if (isSparkVersionGE("3.5")) {
+      assertResult(6)(objectCount)
+      assert(hasCommits)
+    } else {
+      assertResult(5)(objectCount)
+    }
+
+    assert(metadataGlutenExist)
+    assert(metadataBinExist)
+    assert(dataBinExist)
   }
 
   test("test mergetree write with orderby keys / primary keys") {
     spark.sql(s"""
-                 |DROP TABLE IF EXISTS lineitem_mergetree_orderbykey_hdfs;
+                 |DROP TABLE IF EXISTS lineitem_mergetree_orderbykey_s3;
                  |""".stripMargin)
 
     spark.sql(s"""
-                 |CREATE TABLE IF NOT EXISTS lineitem_mergetree_orderbykey_hdfs
+                 |CREATE TABLE IF NOT EXISTS lineitem_mergetree_orderbykey_s3
                  |(
                  | l_orderkey      bigint,
                  | l_partkey       bigint,
@@ -186,14 +245,14 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
                  | l_comment       string
                  |)
                  |USING clickhouse
-                 |TBLPROPERTIES (storage_policy='__hdfs_main',
+                 |TBLPROPERTIES (storage_policy='__s3_main',
                  |               orderByKey='l_shipdate,l_orderkey',
                  |               primaryKey='l_shipdate')
-                 |LOCATION '$HDFS_URL/test/lineitem_mergetree_orderbykey_hdfs'
+                 |LOCATION 's3a://$BUCKET_NAME/lineitem_mergetree_orderbykey_s3'
                  |""".stripMargin)
 
     spark.sql(s"""
-                 | insert into table lineitem_mergetree_orderbykey_hdfs
+                 | insert into table lineitem_mergetree_orderbykey_s3
                  | select * from lineitem
                  |""".stripMargin)
 
@@ -211,7 +270,7 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
          |    avg(l_discount) AS avg_disc,
          |    count(*) AS count_order
          |FROM
-         |    lineitem_mergetree_orderbykey_hdfs
+         |    lineitem_mergetree_orderbykey_s3
          |WHERE
          |    l_shipdate <= date'1998-09-02' - interval 1 day
          |GROUP BY
@@ -252,16 +311,16 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
         assertResult(1)(addFiles.size)
         assertResult(600572)(addFiles.head.rows)
     }
-    spark.sql("drop table lineitem_mergetree_orderbykey_hdfs")
+    spark.sql("drop table lineitem_mergetree_orderbykey_s3")
   }
 
   test("test mergetree write with partition") {
     spark.sql(s"""
-                 |DROP TABLE IF EXISTS lineitem_mergetree_partition_hdfs;
+                 |DROP TABLE IF EXISTS lineitem_mergetree_partition_s3;
                  |""".stripMargin)
 
     spark.sql(s"""
-                 |CREATE TABLE IF NOT EXISTS lineitem_mergetree_partition_hdfs
+                 |CREATE TABLE IF NOT EXISTS lineitem_mergetree_partition_s3
                  |(
                  | l_orderkey      bigint,
                  | l_partkey       bigint,
@@ -282,15 +341,15 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
                  |)
                  |USING clickhouse
                  |PARTITIONED BY (l_returnflag)
-                 |TBLPROPERTIES (storage_policy='__hdfs_main',
+                 |TBLPROPERTIES (storage_policy='__s3_main',
                  |               orderByKey='l_orderkey',
                  |               primaryKey='l_orderkey')
-                 |LOCATION '$HDFS_URL/test/lineitem_mergetree_partition_hdfs'
+                 |LOCATION 's3a://$BUCKET_NAME/lineitem_mergetree_partition_s3'
                  |""".stripMargin)
 
     // dynamic partitions
     spark.sql(s"""
-                 | insert into table lineitem_mergetree_partition_hdfs
+                 | insert into table lineitem_mergetree_partition_s3
                  | select * from lineitem
                  |""".stripMargin)
 
@@ -320,11 +379,11 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
     source.write
       .format("clickhouse")
       .mode(SaveMode.Append)
-      .insertInto("lineitem_mergetree_partition_hdfs")
+      .insertInto("lineitem_mergetree_partition_s3")
 
     // static partition
     spark.sql(s"""
-                 | insert into lineitem_mergetree_partition_hdfs PARTITION (l_returnflag = 'A')
+                 | insert into lineitem_mergetree_partition_s3 PARTITION (l_returnflag = 'A')
                  | (l_shipdate,
                  |  l_orderkey,
                  |  l_partkey,
@@ -373,7 +432,7 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
          |    avg(l_discount) AS avg_disc,
          |    count(*) AS count_order
          |FROM
-         |    lineitem_mergetree_partition_hdfs
+         |    lineitem_mergetree_partition_s3
          |WHERE
          |    l_shipdate <= date'1998-09-02' - interval 1 day
          |GROUP BY
@@ -431,16 +490,17 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
         assertResult(6)(addFiles.size)
         assertResult(750735)(addFiles.map(_.rows).sum)
     }
-    spark.sql("drop table lineitem_mergetree_partition_hdfs")
+    spark.sql("drop table lineitem_mergetree_partition_s3")
+
   }
 
   test("test mergetree write with bucket table") {
     spark.sql(s"""
-                 |DROP TABLE IF EXISTS lineitem_mergetree_bucket_hdfs;
+                 |DROP TABLE IF EXISTS lineitem_mergetree_bucket_s3;
                  |""".stripMargin)
 
     spark.sql(s"""
-                 |CREATE TABLE IF NOT EXISTS lineitem_mergetree_bucket_hdfs
+                 |CREATE TABLE IF NOT EXISTS lineitem_mergetree_bucket_s3
                  |(
                  | l_orderkey      bigint,
                  | l_partkey       bigint,
@@ -463,12 +523,12 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
                  |PARTITIONED BY (l_returnflag)
                  |CLUSTERED BY (l_orderkey)
                  |${if (sparkVersion.equals("3.2")) "" else "SORTED BY (l_partkey)"} INTO 4 BUCKETS
-                 |LOCATION '$HDFS_URL/test/lineitem_mergetree_bucket_hdfs'
-                 |TBLPROPERTIES (storage_policy='__hdfs_main')
+                 |LOCATION 's3a://$BUCKET_NAME/lineitem_mergetree_bucket_s3'
+                 |TBLPROPERTIES (storage_policy='__s3_main')
                  |""".stripMargin)
 
     spark.sql(s"""
-                 | insert into table lineitem_mergetree_bucket_hdfs
+                 | insert into table lineitem_mergetree_bucket_s3
                  | select * from lineitem
                  |""".stripMargin)
 
@@ -486,7 +546,7 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
          |    avg(l_discount) AS avg_disc,
          |    count(*) AS count_order
          |FROM
-         |    lineitem_mergetree_bucket_hdfs
+         |    lineitem_mergetree_bucket_s3
          |WHERE
          |    l_shipdate <= date'1998-09-02' - interval 1 day
          |GROUP BY
@@ -532,11 +592,11 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
         assertResult(12)(addFiles.size)
         assertResult(600572)(addFiles.map(_.rows).sum)
     }
-    spark.sql("drop table lineitem_mergetree_bucket_hdfs")
+    spark.sql("drop table lineitem_mergetree_bucket_s3")
   }
 
   test("test mergetree write with the path based") {
-    val dataPath = s"$HDFS_URL/test/lineitem_mergetree_bucket_hdfs"
+    val dataPath = s"s3a://$BUCKET_NAME/lineitem_mergetree_bucket_s3"
 
     val sourceDF = spark.sql(s"""
                                 |select * from lineitem
@@ -550,7 +610,7 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
       .option("clickhouse.primaryKey", "l_orderkey")
       .option("clickhouse.numBuckets", "4")
       .option("clickhouse.bucketColumnNames", "l_orderkey")
-      .option("clickhouse.storage_policy", "__hdfs_main")
+      .option("clickhouse.storage_policy", "__s3_main")
       .save(dataPath)
 
     val sqlStr =
@@ -618,14 +678,13 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
   }
 
   test("test mergetree insert with optimize basic") {
-    val tableName = "lineitem_mergetree_insert_optimize_basic_hdfs"
-    val dataPath = s"$HDFS_URL/test/$tableName"
+    val tableName = "lineitem_mergetree_insert_optimize_basic_s3"
+    val dataPath = s"s3a://$BUCKET_NAME/$tableName"
 
     withSQLConf(
       "spark.databricks.delta.optimize.minFileSize" -> "200000000",
-      "spark.gluten.sql.columnar.backend.ch.runtime_settings.mergetree.merge_after_insert" -> "true",
       "spark.gluten.sql.columnar.backend.ch.runtime_settings.mergetree.insert_without_local_storage" -> "true",
-      "spark.gluten.sql.columnar.backend.ch.runtime_settings.min_insert_block_size_rows" -> "10000"
+      "spark.gluten.sql.columnar.backend.ch.runtime_settings.mergetree.merge_after_insert" -> "true"
     ) {
       spark.sql(s"""
                    |DROP TABLE IF EXISTS $tableName;
@@ -635,26 +694,129 @@ class GlutenClickHouseMergeTreeWriteOnHDFSSuite
                    |CREATE TABLE IF NOT EXISTS $tableName
                    |USING clickhouse
                    |LOCATION '$dataPath'
-                   |TBLPROPERTIES (storage_policy='__hdfs_main')
                    | as select * from lineitem
                    |""".stripMargin)
 
       val ret = spark.sql(s"select count(*) from $tableName").collect()
       assertResult(600572)(ret.apply(0).get(0))
-      val conf = new Configuration
-      conf.set("fs.defaultFS", HDFS_URL)
-      val fs = FileSystem.get(conf)
+      assert(
+        !new File(s"$CH_DEFAULT_STORAGE_DIR/lineitem_mergetree_insert_optimize_basic").exists())
+    }
+  }
 
-      eventually(timeout(60.seconds), interval(2.seconds)) {
-        val it = fs.listFiles(new Path(dataPath), true)
-        var files = 0
-        while (it.hasNext) {
-          it.next()
-          files += 1
-        }
-        assertResult(4)(files)
+  test("test mergetree with primary keys pruning by driver") {
+    val tableName = "lineitem_mergetree_pk_pruning_by_driver_s3"
+    val dataPath = s"s3a://$BUCKET_NAME/$tableName"
+    spark.sql(s"""
+                 |DROP TABLE IF EXISTS $tableName;
+                 |""".stripMargin)
+
+    spark.sql(s"""
+                 |CREATE TABLE IF NOT EXISTS $tableName
+                 |(
+                 | l_orderkey      bigint,
+                 | l_partkey       bigint,
+                 | l_suppkey       bigint,
+                 | l_linenumber    bigint,
+                 | l_quantity      double,
+                 | l_extendedprice double,
+                 | l_discount      double,
+                 | l_tax           double,
+                 | l_returnflag    string,
+                 | l_linestatus    string,
+                 | l_shipdate      date,
+                 | l_commitdate    date,
+                 | l_receiptdate   date,
+                 | l_shipinstruct  string,
+                 | l_shipmode      string,
+                 | l_comment       string
+                 |)
+                 |USING clickhouse
+                 |TBLPROPERTIES (storage_policy='__s3_main', orderByKey='l_shipdate')
+                 |LOCATION '$dataPath'
+                 |""".stripMargin)
+
+    spark.sql(s"""
+                 | insert into table $tableName
+                 | select * from lineitem
+                 |""".stripMargin)
+
+    FileUtils.forceDelete(new File(S3_METADATA_PATH))
+
+    val sqlStr =
+      s"""
+         |SELECT
+         |    sum(l_extendedprice * l_discount) AS revenue
+         |FROM
+         |    $tableName
+         |WHERE
+         |    l_shipdate >= date'1994-01-01'
+         |    AND l_shipdate < date'1994-01-01' + interval 1 year
+         |    AND l_discount BETWEEN 0.06 - 0.01 AND 0.06 + 0.01
+         |    AND l_quantity < 24
+         |""".stripMargin
+
+    withSQLConf(
+      "spark.gluten.sql.columnar.backend.ch.runtime_settings.enabled_driver_filter_mergetree_index" -> "true") {
+      runTPCHQueryBySQL(6, sqlStr) {
+        df =>
+          val scanExec = collect(df.queryExecution.executedPlan) {
+            case f: FileSourceScanExecTransformer => f
+          }
+          assertResult(1)(scanExec.size)
+
+          val mergetreeScan = scanExec.head
+          assert(mergetreeScan.nodeName.startsWith("Scan mergetree"))
+
+          val plans = collect(df.queryExecution.executedPlan) {
+            case scanExec: BasicScanExecTransformer => scanExec
+          }
+          assertResult(1)(plans.size)
+          assertResult(1)(plans.head.getSplitInfos.size)
       }
     }
   }
+
+  test("GLUTEN-6750: Optimize error if file metadata not exist") {
+    spark.sql(s"""
+                 |DROP TABLE IF EXISTS lineitem_mergetree_bucket_s3;
+                 |""".stripMargin)
+
+    spark.sql(s"""
+                 |CREATE TABLE IF NOT EXISTS lineitem_mergetree_bucket_s3
+                 |(
+                 | l_orderkey      bigint,
+                 | l_partkey       bigint,
+                 | l_suppkey       bigint,
+                 | l_linenumber    bigint,
+                 | l_quantity      double,
+                 | l_extendedprice double,
+                 | l_discount      double,
+                 | l_tax           double,
+                 | l_returnflag    string,
+                 | l_linestatus    string,
+                 | l_shipdate      date,
+                 | l_commitdate    date,
+                 | l_receiptdate   date,
+                 | l_shipinstruct  string,
+                 | l_shipmode      string,
+                 | l_comment       string
+                 |)
+                 |USING clickhouse
+                 |PARTITIONED BY (l_returnflag)
+                 |CLUSTERED BY (l_orderkey)
+                 |${if (sparkVersion.equals("3.2")) "" else "SORTED BY (l_partkey)"} INTO 4 BUCKETS
+                 |LOCATION 's3a://$BUCKET_NAME/lineitem_mergetree_bucket_s3'
+                 |TBLPROPERTIES (storage_policy='__s3_main')
+                 |""".stripMargin)
+
+    spark.sql(s"""
+                 | insert into table lineitem_mergetree_bucket_s3
+                 | select /*+ REPARTITION(3) */ * from lineitem
+                 |""".stripMargin)
+
+    FileUtils.deleteDirectory(new File(S3_METADATA_PATH))
+    spark.sql("optimize lineitem_mergetree_bucket_s3")
+    spark.sql("drop table lineitem_mergetree_bucket_s3")
+  }
 }
-// scalastyle:off line.size.limit
